joyce---chatbot-zur-studienberatung-master/middleware/requirements.txt
asn1crypto==0.24.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
dialogflow==0.5.1
Flask==1.0.2
Flask-Cors==3.0.7
google-api-core==1.7.0
google-auth==1.6.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
gunicorn==19.9.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.7
redis==2.10.6
requests==2.18.4
rsa==4.0
six==1.12.0
urllib3==1.22
Werkzeug==0.14.1
